third_party/blink/web_tests/platform/linux/virtual/webnn-service-with-gpu/external/wpt/webnn/conformance_tests/tile.https.any_gpu-expected.txt

This is a testharness.js-based test.
[FAIL] tile float16 1D constant tensor
  assert_less_than_equal: assert_array_approx_equals_ulp: test tile float16 actual 0 should be close enough to expected 1 by ULP distance: expected a number less than or equal to 0 but got 15360
[FAIL] tile float16 2D tensor
  assert_less_than_equal: assert_array_approx_equals_ulp: test tile float16 actual 0 should be close enough to expected 1 by ULP distance: expected a number less than or equal to 0 but got 15360
[FAIL] tile uint32 2D tensor
  promise_test: Unhandled rejection with value: object "TypeError: Unsupported data type, input 'input' data type uint32 must be one of [float32,float16,int32,int64,uint8]."
Harness: the test ran to completion.
